Move session module (non-breaking)
This moves the sesson module out of types and into
the location indicated in commit 20b26a5.

src/types/session.rs -> src/session/mod.rs
src/types/mod.rs::SessionFlags -> src/session/flags.rs
