+++ cryptoki/src/functions/object_management.rs
@@ -5,7 +5,7 @@
 use crate::get_pkcs11;
 use crate::types::function::{Rv, RvError};
 use crate::types::object::{Attribute, AttributeInfo, AttributeType, ObjectHandle};
-use crate::types::session::Session;
+use crate::session::Session;
 use crate::Result;
 use cryptoki_sys::*;
 use std::collections::HashMap;
