+++ cryptoki/src/functions/random.rs
@@ -4,7 +4,7 @@
 
 use crate::get_pkcs11;
 use crate::types::function::Rv;
-use crate::types::session::Session;
+use crate::session::Session;
 use crate::Result;
 use std::convert::TryInto;
