+++ cryptoki/src/functions/session_management.rs
@@ -4,9 +4,8 @@
 
 use crate::get_pkcs11;
 use crate::types::function::Rv;
-use crate::types::session::{Session, SessionInfo, UserType};
+use crate::session::{Session, SessionInfo, SessionFlags, UserType};
 use crate::types::slot_token::Slot;
-use crate::types::SessionFlags;
 use crate::Pkcs11;
 use crate::Result;
 use cryptoki_sys::CK_SESSION_INFO;
