+++ cryptoki/src/functions/signing_macing.rs
@@ -6,7 +6,7 @@ use crate::get_pkcs11;
 use crate::types::function::Rv;
 use crate::mechanism::Mechanism;
 use crate::types::object::ObjectHandle;
-use crate::types::session::Session;
+use crate::session::Session;
 use crate::Result;
 use cryptoki_sys::*;
 use std::convert::TryInto;
